+++ rayTracer/src/world.h
@@ -1,19 +1,18 @@
 #pragma once
 
 #include "geometry.h"
+#include "objectLoader.h"
 #include "pointLight.h"
 #include "ray.h"
 #include "shape.h"
 #include "sphere.h"
-#include "objectLoader.h"
 #include <glm/glm.hpp>
 #include <iostream>
 #include <memory>
-#include <vector>
 #include <typeinfo>
+#include <vector>
 
-class World
-{
+class World {
 private:
   /* data */
   std::unique_ptr<ObjectLoader> objectLoader;
@@ -23,11 +22,11 @@ public:
   ~World();
 
   std::shared_ptr<Camera> loadFromFile(const std::string &fileName);
-  
+
   void addShape(std::shared_ptr<Shape> &shape_ptr);
   void addLight(std::shared_ptr<PointLight> &light);
 
-  bool isShadowed(glm::vec4 point);
+  bool isShadowed(glm::dvec4 point);
   std::vector<Geometry::Intersection<Shape>> intersectRay(Ray &ray);
 
   std::vector<std::shared_ptr<Shape>> shapes;
